Extract new component into its own module
To keep things clean, open to feedback on naming
